Optimise handling of aliases and anchors.

See: https://github.com/tlsa/libcyaml/issues/129

| Branch                  | `VARIANT=debug` | `VARIANT=release` |
| ----------------------- | --------------- | ----------------- |
| `master`                | 8m 51.088s      | 3m 12.613s        |
| `tlsa/optimise-anchors` | 2.773s          | 1.843s            |